// src/app/models/requirement.model.ts
export interface Requirement {
  requirementId: string;
  requirementNo: string;
  userId: string;
  buyerNo: string;
  buyerName: string;
  title: string;
  productImage: string;
  productDescription: string;
  status: string;
  statusEnum: number;
  createdDate: string;
  quantity: number;
  unit: string;
  city: string;
  quotes: number;
  totalPriceEstimation: number;
  preferInIndiaProducts: boolean;
  expectedDeliveryDate: string;
  category: string;
  subCategories: string[];
}

export interface RequirementDetail {
  requirementId: string;
  requirementNo: string;
  userId: string;
  title: string;
  buyerName: string;
  buyerNo: string;
  productImage: string;
  productDescription: string;
  quantity: number;
  unit: string;
  status: string;
  totalPriceEstimation: number;
  preferInIndiaProducts: boolean;
  pickupProductDirectly: boolean;
  deliveryLocationPinCode: string;
  preferredSourceOfSupply: string;
  expectedDeliveryDate: string;
  expectedDeliveryDateStr: string;
  billingAddressName: string;
  billingAddressBuilding: string;
  billingAddressStreet: string;
  billingAddressCity: string;
  billingAddressState: string;
  billingAddressPinCode: string;
  shippingAddressName: string;
  shippingAddressBuilding: string;
  shippingAddressStreet: string;
  shippingAddressCity: string;
  shippingAddressState: string;
  shippingAddressPinCode: string;
  shippingAddressLandmark: string;
  needInsuranceCoverage: boolean;
  category: string;
  subCategories: string[];
  statusEnum: number;
  isDeleted: boolean;
  createdDate: string;
  canRevealContact: boolean;
  receivedQuotes: [];
  buyerContact?: any;
  isBuyerContactRevealed: boolean;
  isReseller: boolean;
  agreeGSTc: boolean;
  gstin: string;
}

export interface RequirementRequest {
  title: string;
  totalPriceEstimation: number;
  category: string;
  subCategories: string[];
  productDescription: string;
  quantity: number;
  unit: string;
  productImage: string;
  gstin: string;
  expectedDeliveryDateStr: string;
  preferredSourceOfSupply: string;
  isReseller: boolean;
  preferInIndiaProducts: boolean;
  pickupProductDirectly: boolean;
  needInsuranceCoverage: boolean;
  agreeGSTc: boolean;
  deliveryLocationPinCode: string;
  billingAddressBuilding: string;
  billingAddressCountry: number;
  billingAddressCity: string;
  billingAddressName: string;
  billingAddressState: number;
  billingAddressStreet: string;
  billingAddressPinCode: string;
  billingAddressLandmark: string;
  shippingAddressStreet: string;
  shippingAddressCountry: number;
  shippingAddressState: number;
  shippingAddressPinCode: string;
  shippingAddressName: string;
  shippingAddressCity: string;
  shippingAddressBuilding: string;
  shippingAddressLandmark: string;
}

export interface RequirementSeller {
  requirementId: string;
  requirementNo: string;
  userId: string;
  buyerNo: string;
  buyerName: string;
  title: string;
  productImage: string;
  productDescription: string;
  status: string;
  statusEnum: number;
  createdDate: string;
  quantity: number;
  unit: string;
  city: string;
  totalPriceEstimation: number;
  preferInIndiaProducts: boolean;
  expectedDeliveryDate: string;
  category: string;
  subCategories: string[];
  pickupProductDirectly: boolean;
  needInsuranceCoverage: boolean;
}